+++ frontend/appflowy_flutter/lib/features/share/data/models/share_role.dart
@@ -1,0 +1,14 @@
+/// The role a user can have on a shared page.
+enum ShareRole {
+  /// Can view the page only.
+  readOnly,
+
+  /// Can read and comment on the page.
+  readAndComment,
+
+  /// Can read and write to the page.
+  readAndWrite,
+
+  /// Full access (edit, share, remove, etc.) and can add new users.
+  fullAccess,
+}
