+++ frontend/appflowy_flutter/lib/features/share/data/models/shared_user.dart
@@ -1,0 +1,16 @@
+import 'package:appflowy/features/share/data/models/share_role.dart';
+
+/// Represents a user with a role on a shared page.
+class SharedUser {
+  SharedUser({
+    required this.email,
+    required this.name,
+    required this.role,
+    this.avatarUrl,
+  });
+
+  final String email;
+  final String name;
+  final ShareRole role;
+  final String? avatarUrl;
+}
