+++ frontend/appflowy_flutter/lib/features/share/data/repositories/share_repository.dart
@@ -1,18 +1,12 @@
+import 'package:appflowy/features/share/data/models/share_role.dart';
+import 'package:appflowy/features/share/data/models/shared_user.dart';
 import 'package:appflowy_backend/protobuf/flowy-error/errors.pb.dart';
 import 'package:appflowy_result/appflowy_result.dart';
 
-/// Represents a user with a role on a shared page.
-class SharedUser {
-  SharedUser({
-    required this.email,
-    required this.role,
-  });
-
-  final String email;
-  final ShareRole role;
-}
-
 /// Abstract repository for sharing pages with users.
+///
+/// For example, we're using rust events now, but we can still use the http api
+/// for the future.
 abstract class ShareRepository {
   /// Gets the list of users and their roles for a shared page.
   Future<FlowyResult<List<SharedUser>, FlowyError>> getUsersInSharedPage({
@@ -33,19 +27,4 @@ abstract class ShareRepository {
     required ShareRole role,
     required List<String> emails,
   });
-}
-
-/// The role a user can have on a shared page.
-enum ShareRole {
-  /// Can view the page only.
-  readOnly,
-
-  /// Can read and comment on the page.
-  readAndComment,
-
-  /// Can read and write to the page.
-  readAndWrite,
-
-  /// Full access (edit, share, remove, etc.) and can add new users.
-  fullAccess,
 }
